Extract post-walk action reconciler
